Code quality (#104)
* Abstracted display message

* Added observe share

* More message abstraction

* Added another empty
