+++ app/src/main/java/com/koalatea/thehollidayinn/softwareengineeringdaily/subscription/PaymentFragment.kt
@@ -105,7 +105,7 @@ class PaymentFragment : Fragment() {
         if (context is OnFragmentInteractionListener) {
             mListener = context
         } else {
-            throw RuntimeException(context!!.toString() + " must implement OnFragmentInteractionListener")
+            throw RuntimeException(context!!.toString() + " must implement PaymentFragment.OnFragmentInteractionListener")
         }
     }
