+++ app/src/main/java/com/koalatea/thehollidayinn/softwareengineeringdaily/subscription/PlanFragment.kt
@@ -41,7 +41,7 @@ class PlanFragment : Fragment() {
         if (context is OnFragmentInteractionListener) {
             mListener = context
         } else {
-            throw RuntimeException(context!!.toString() + " must implement OnFragmentInteractionListener")
+            throw RuntimeException(context!!.toString() + " must implement PlanFragment.OnFragmentInteractionListener")
         }
     }
